Make LogEntry public
This is useful for users who have their own http client implementation
